requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
Bottleneck==1.3.7
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
dataclasses-json==0.6.6
debugpy==1.6.7
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.0
executing==2.0.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
ipykernel==6.29.3
ipython==8.24.0
jedi==0.19.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.1
jupyter_core==5.5.0
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-openai==0.1.6
langchain-text-splitters==0.0.1
langsmith==0.1.57
marshmallow==3.21.2
matplotlib-inline==0.1.7
multidict==6.0.5
mypy-extensions==1.0.0
nest_asyncio==1.6.0
nltk==3.8.1
numexpr==2.8.7
numpy==1.26.4
openai==1.28.1
orjson==3.10.3
packaging==23.2
pandas==2.2.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pip==24.0
platformdirs==4.2.1
praw==7.7.1
prawcore==2.4.0
prompt-toolkit==3.0.42
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
python-dateutil==2.9.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
regex==2024.5.10
requests==2.31.0
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
stack-data==0.6.2
tenacity==8.3.0
tiktoken==0.6.0
tornado==6.3.3
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.11.0
typing-inspect==0.9.0
tzdata==2023.3
update-checker==0.18.0
urllib3==2.2.1
wcwidth==0.2.13
websocket-client==1.8.0
wheel==0.43.0
yarl==1.9.4
zipp==3.17.0
